+++ mitra-models/src/posts/queries.rs
@@ -1442,18 +1442,25 @@ pub async fn delete_post(
     })
 }
 
-pub async fn get_local_post_count(
+pub async fn get_post_count(
     db_client: &impl DatabaseClient,
+    only_local: bool,
 ) -> Result<i64, DatabaseError> {
-    let row = db_client.query_one(
+    let mut condition =
+        "post.in_reply_to_id IS NULL AND post.repost_of_id IS NULL".to_string();
+    if only_local {
+        condition.push_str(" AND actor_profile.actor_json IS NULL");
+    };
+    let statement = format!(
         "
         SELECT count(post)
         FROM post
-        JOIN user_account ON (post.author_id = user_account.id)
-        WHERE post.in_reply_to_id IS NULL AND post.repost_of_id IS NULL
+        JOIN actor_profile ON (post.author_id = actor_profile.id)
+        WHERE {condition}
         ",
-        &[],
-    ).await?;
+        condition=condition,
+    );
+    let row = db_client.query_one(&statement, &[]).await?;
     let count = row.try_get("count")?;
     Ok(count)
 }