+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Show total number of posts in instance report.
+
 ### Changed
 
 - Allow to import objects with type `Proposal` as posts.